WorkingScopes

WorkingScopes.jl has no utility; it is merely a package of working examples for demonstration and experiments to understand Scope of Variables as well as Metaprogramming in julia programming.

Documentation for WorkingScopes.

WorkingScopes.sayhello3

WorkingScopes.sayhello3

WorkingScopes.@sayhello3

WorkingScopes.@sayhello3

WorkingScopes.@sayhello3

WorkingScopes.@sayhello3

WorkingScopes.sayhello3 — Method

sayhello3(name) = "Hello world! $name", an example to understand metaprogramming.

WorkingScopes.@sayhello3 — Macro

macro sayhello3(ex::Expr)
    return Expr(:block, :(sayhello3($(last(ex.args)))))
end

Example: variable name is not defined in macro's scope

name = "Bruce Willey"

# output
"Bruce Willey"

julia> WorkingScopes.@sayhello3 $name
ERROR: UndefVarError: `name` not defined

Explain

name is passed as expression in this case, and it is not defined since macro dispatch is not based on the AST evaluated at runtime. Thus an error of UndefVarError is raised.

See Macros and dispatch for more information.

Example: a creative application

WorkingScopes.@sayhello3 name = "Bruce Willey"

# output
"Hello world! Bruce Willey"

Explain

In this example, the last argument of the expression name = "Bruce Willey" is "Bruce Willey".

Example 3: Use @eval if you want name be evaluated

@eval WorkingScopes.@sayhello3 $name

# output
"Hello world! Bruce Willey"

Explain

@eval evalute the expression @sayhello3 $name at runtime.

Thus, the argument for @sayhello3 is evaluated as String, and @sayhello3(str::String) is the dispatched method.

WorkingScopes.@sayhello3 — Macro

macro sayhello3(str::String)
    return Expr(:block, :(sayhello3($str)))
end

Example

WorkingScopes.@sayhello3 "Bruce Willey"

# output
"Hello world! Bruce Willey"

WorkingScopes.@sayhello3 — Macro

macro sayhello3(symb::Symbol)
    return Expr(:block, :(sayhello3($(string(symb)))))
end

Example: A single the runtime variable is rendered as Symbol

name = "Bruce Willey"
WorkingScopes.@sayhello3 name

# output
"Hello world! name"