Macro: create expression to be executed in its surrounded scope

The use of macro is confusing. Thus, I made an exemplary @sayhello3 to help me understand.

macro returns expression, and the ex::Expr is executed in the scope just like @eval(mod::Module, ex) (mod can be Main).

ex::Expr uses the surrounded scope (lexical scope):

macro f45()
    x = 5
    return :(f(45))
end

@f45 (macro with 1 method)

f = sin
@f45

0.8509035245341184

f = tan
@f45

1.6197751905438615

Variable defined in expression won't contaminate the surrounded scope,

julia> @f451.6197751905438615
julia> xERROR: UndefVarError: `x` not defined

unless

module Hello
    macro f45brutal()
        Hello.x = 5
        return :(f(45))
    end
    Hello.x = 4.99
end

Main.Hello

julia> Hello.x4.99
julia> Hello.@f45brutalERROR: UndefVarError: `f` not defined
julia> Hello.x5

Error occurred in Hello.@f45brutal because Hello.f is not defined. Noted that this is true for Main scope; yuo can redefine variables/functions under the Main scope this way.

Reference

Emma Boudreau - Metaprogramming in Julia: A Full Overview

Tips

Tips

Multiline quote ... end is in fact Expr(:block, ex1, ex2, ...)

esc escape ex::Expr from macro expansion; it is used to "violate" macro hygiene when necessary. See Metaprogramming/#Hygiene.

macroexpand and @macroexpand is very useful in debugging. See metaprogramming/#Hold-up:-why-macros?

Use macro or @evaluate an expression

It is generally safe to use macro defined in a clean module.

module HelloWorld
    f = sin

    macro definex(x)
        return quote
            HelloWorld.x = $x
            HelloWorld.sinx = f($x)
        end
    end

    HelloWorld.x = 5.99
    HelloWorld.sinx = f(x)
end

Main.HelloWorld

HelloWorld.x

5.99

HelloWorld.sinx

-0.2890030703793611

x = 5
f = cos
f(x)

0.28366218546322625

HelloWorld.@definex 0
HelloWorld.x

0

HelloWorld.sinx

0.0

x

5

Tip

Expression returned by macro is executed under the scope that the macro is defined.

Noted that in @definex use f defined in module HelloWorld ... end

The same result/utility can be achieved with function using @eval:

module HelloWorld
    f = sin

    function definex(x)
        ex = quote
            HelloWorld.x = $x
            HelloWorld.sinx = f($x)
        end
        @eval $ex
    end

    HelloWorld.x = 5.99
    HelloWorld.sinx = f(x)
end
HelloWorld.sinx

-0.2890030703793611

HelloWorld.definex(0)
x = 5
HelloWorld.x

0

HelloWorld.sinx

0.0

Be ware of the difference between @eval ex and @eval $ex

f = sin
x = 0
ex = quote
    x = $x
    sinx = f($x)
end

quote
    #= sayhello.md:159 =#
    x = 0
    #= sayhello.md:160 =#
    sinx = f(0)
end

julia> @eval exquote
    #= sayhello.md:159 =#
    x = 0
    #= sayhello.md:160 =#
    sinx = f(0)
end
julia> @macroexpand @eval ex:(Core.eval(Main, :ex))
julia> sinxERROR: UndefVarError: `sinx` not defined

The ERROR: UndefVarError: 'sinx' not defined occurred since ex is passed to @eval as Symbol; thus, nothing executed. Noted that ERROR: UndefVarError: 'ex' not defined won't be raised when ex is not defined!

julia> @eval $ex0.0
julia> @macroexpand @eval $ex:(Core.eval(Main, ex))
julia> sinx0.0

Example

Say Hello

the say-hello function

WorkingScopes.sayhello3 — Function

sayhello3(name) = "Hello world! $name", an example to understand metaprogramming.

The say-hello macro built upon

WorkingScopes.@sayhello3 — Macro

macro sayhello3(str::String)
    return Expr(:block, :(sayhello3($str)))
end

Example

WorkingScopes.@sayhello3 "Bruce Willey"

# output
"Hello world! Bruce Willey"

macro sayhello3(ex::Expr)
    return Expr(:block, :(sayhello3($(last(ex.args)))))
end

Example: variable name is not defined in macro's scope

name = "Bruce Willey"

# output
"Bruce Willey"

julia> WorkingScopes.@sayhello3 $name
ERROR: UndefVarError: `name` not defined

Explain

name is passed as expression in this case, and it is not defined since macro dispatch is not based on the AST evaluated at runtime. Thus an error of UndefVarError is raised.

See Macros and dispatch for more information.

Example: a creative application

WorkingScopes.@sayhello3 name = "Bruce Willey"

# output
"Hello world! Bruce Willey"

Explain

In this example, the last argument of the expression name = "Bruce Willey" is "Bruce Willey".

Example 3: Use @eval if you want name be evaluated

@eval WorkingScopes.@sayhello3 $name

# output
"Hello world! Bruce Willey"

Explain

@eval evalute the expression @sayhello3 $name at runtime.

Thus, the argument for @sayhello3 is evaluated as String, and @sayhello3(str::String) is the dispatched method.

macro sayhello3(symb::Symbol)
    return Expr(:block, :(sayhello3($(string(symb)))))
end

Example: A single the runtime variable is rendered as Symbol

name = "Bruce Willey"
WorkingScopes.@sayhello3 name

# output
"Hello world! name"